refactor: 프레임워크 클래스 패키지 이동
