+++ app/src/main/java/com/techcourse/DispatcherServletInitializer.java
@@ -4,6 +4,7 @@ import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import web.org.springframework.web.WebApplicationInitializer;
+import webmvc.org.springframework.web.servlet.mvc.DispatcherServlet;
 import webmvc.org.springframework.web.servlet.mvc.asis.MyControllerHandlerAdapter;
 import webmvc.org.springframework.web.servlet.mvc.support.HandlerNotFoundExceptionResolver;
 import webmvc.org.springframework.web.servlet.mvc.tobe.AnnotationHandlerMapping;
