+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/support/HandlerAdapters.java
@@ -1,8 +1,8 @@
-package com.techcourse.support;
+package webmvc.org.springframework.web.servlet.mvc.support;
 
-import jakarta.annotation.Nonnull;
 import java.util.ArrayList;
 import java.util.List;
+import javax.annotation.Nonnull;
 import webmvc.org.springframework.web.servlet.mvc.HandlerAdapter;
 import webmvc.org.springframework.web.servlet.mvc.exception.HandlerAdapterNotFoundException;
