+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/support/HandlerExceptionResolvers.java
@@ -1,10 +1,10 @@
-package com.techcourse.support;
+package webmvc.org.springframework.web.servlet.mvc.support;
 
-import jakarta.annotation.Nonnull;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.HashMap;
 import java.util.Map;
+import javax.annotation.Nonnull;
 import webmvc.org.springframework.web.servlet.ModelAndView;
 import webmvc.org.springframework.web.servlet.mvc.HandlerExceptionResolver;
