+++ mvc/src/test/java/webmvc/org/springframework/web/servlet/mvc/support/HandlerAdaptersTest.java
@@ -1,10 +1,11 @@
-package com.techcourse.support;
+package webmvc.org.springframework.web.servlet.mvc.support;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
+import org.junit.jupiter.api.Test;
 import webmvc.org.springframework.web.servlet.mvc.exception.HandlerAdapterNotFoundException;
 import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerExecution;
 import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerExecutionHandlerAdapter;
